Clarify when rebin needs scaling
It's only necessary if the bins are already normalized by the bin width.
For the jet hadron correlation analysis, we normalize by the bin width
later, so it's not necessary. In contrast, the RM hists are already
normalized by the bin width, so we scale to ensure that we don't lose
that physical meaning