Alternative RPF width work in progress code
This would extract the fit result values from the RP fit and convert
them to width fit result format. However, this can be rather
complicated to get every value correct, and it isn't of a huge benefit.
The better part of the signal fit is better constrain the v_n terms, so
the width values are not necessarily so important