Cache I/O for ROOT and YAML files
This provides a huge speed up in the analysis. It takes advantage of
`joblib.Memory` to handle the caching. This provides similar performance
to implementing this by hand, but it is much simpler and I don't need to
worry about the code that does the heavily lifting for the cache.