Trivial dependency update
Note that this isn't only the packages required to perform the
analysis - they are all the packages in the virtualenv. So there may be
some that are not strictly required. This commit removes one of those
examples because I noticed it, but it's generally not worth filtering
them out because they should cause no harm